move test-utils.ts to test/shared/visu
# Conflicts:
#	test/e2e/diagram.navigation.fit.test.ts
