+++ test/e2e/bpmn.rendering.test.ts
@@ -16,7 +16,7 @@ limitations under the License.
 
 import 'jest-playwright-preset';
 import type { Page } from 'playwright';
-import { getBpmnDiagramNames } from './helpers/test-utils';
+import { getBpmnDiagramNames } from '@test/shared/visu/test-utils';
 import type { StyleOptions } from '@test/shared/visu/bpmn-page-utils';
 import { AvailableTestPages, PageTester } from '@test/shared/visu/bpmn-page-utils';
 import type { ImageSnapshotThresholdConfig } from './helpers/visu/image-snapshot-config';
