+++ test/e2e/diagram.navigation.fit.test.ts
@@ -19,7 +19,7 @@ import 'jest-playwright-preset';
 import { join } from 'node:path';
 import type { Page } from 'playwright';
 import { FitType } from '@lib/component/options';
-import { getBpmnDiagramNames } from './helpers/test-utils';
+import { getBpmnDiagramNames } from '@test/shared/visu/test-utils';
 import { AvailableTestPages, PageTester } from '@test/shared/visu/bpmn-page-utils';
 import type { ImageSnapshotThresholdConfig } from './helpers/visu/image-snapshot-config';
 import { ImageSnapshotConfigurator, MultiBrowserImageSnapshotThresholds } from './helpers/visu/image-snapshot-config';
