+++ test/performance/bpmn.navigation.performance.test.ts
@@ -16,7 +16,7 @@ limitations under the License.
 
 import * as fs from 'node:fs';
 import type { Page } from 'playwright';
-import { delay } from '../e2e/helpers/test-utils';
+import { delay } from '@test/shared/visu/test-utils';
 import type { Point } from '@test/shared/visu/bpmn-page-utils';
 import { AvailableTestPages, PageTester } from '@test/shared/visu/bpmn-page-utils';
 import { ChromiumMetricsCollector } from './helpers/metrics-chromium';
